crypto.[md5/rc4/sha1]: remove extraneous casts

+++ vlib/crypto/sha1/sha1.v
@@ -48,7 +48,7 @@ fn (d mut Digest) reset() {
 	d.h[3] = u32(Init3)
 	d.h[4] = u32(Init4)
 	d.nx = 0
-	d.len = u64(0)
+	d.len = 0
 }
 
 // new returns a new Digest (implementing hash.Hash) computing the SHA1 checksum.
@@ -115,7 +115,7 @@ fn (d mut Digest) checksum() []byte {
 	}
 
 	// Length in bits.
-	len <<= u64(3)
+	len <<= 3
 	binary.big_endian_put_u64(mut tmp, len)
 	d.write(tmp.left(8))
